Mv constants to own file
Rename constant to constants
